+++ client/web/src/gettingStartedTour/GettingStartedTourInfo.tsx
@@ -6,8 +6,15 @@ export const GETTING_STARTED_TOUR_MARKER = 'getting-started-tour-info-marker'
 
 interface GettingStartedTourInfoProps {
     className?: string
+    isSourcegraphDotCom: boolean
 }
 
-export const GettingStartedTourInfo: React.FunctionComponent<GettingStartedTourInfoProps> = ({ className }) => (
-    <div className={classNames(GETTING_STARTED_TOUR_MARKER, className)} />
-)
+export const GettingStartedTourInfo: React.FunctionComponent<GettingStartedTourInfoProps> = ({
+    className,
+    isSourcegraphDotCom,
+}) => {
+    if (!isSourcegraphDotCom) {
+        return null
+    }
+    return <div className={classNames(GETTING_STARTED_TOUR_MARKER, className)} />
+}