+++ client/web/src/gettingStartedTour/components/LinkOrAnchor.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Link } from '@sourcegraph/wildcard'
+import { isExternalURL } from '../utils'
+
+interface LinkOrAnchorProps {
+    href: string
+    className?: string
+    onClick?: (event: React.MouseEvent<HTMLAnchorElement>) => void
+}
+export const LinkOrAnchor: React.FunctionComponent<LinkOrAnchorProps> = ({ href, children, ...props }) => (
+    <Link to={href} {...props} {...(isExternalURL(href) && { target: '_blank', rel: 'noopener noreferrer' })}>
+        {children}
+    </Link>
+)