+++ client/web/src/repo/RepositoryFileTreePage.tsx
@@ -54,8 +54,6 @@ export const RepositoryFileTreePage: React.FunctionComponent<RepositoryFileTreeP
 
     const mode = getModeFromPath(filePath)
 
-    const showGettingStartedTour = context.isSourcegraphDotCom && !context.authenticatedUser
-
     // Redirect OpenGrok-style line number hashes (#123, #123-321) to query parameter (?L123, ?L123-321)
     const hashLineNumberMatch = window.location.hash.match(/^#?(\d+)(-\d+)?$/)
     if (objectType === 'blob' && hashLineNumberMatch) {
@@ -105,13 +103,12 @@ export const RepositoryFileTreePage: React.FunctionComponent<RepositoryFileTreeP
                 className="repo-revision-container__sidebar"
                 isDir={objectType === 'tree'}
                 defaultBranch={defaultBranch || 'HEAD'}
-                showGettingStartedTour={showGettingStartedTour}
             />
             {!hideRepoRevisionContent && (
                 // Add `.blob-status-bar__container` because this is the
                 // lowest common ancestor of Blob and the absolutely-positioned Blob status bar
                 <BlobStatusBarContainer>
-                    {showGettingStartedTour && <GettingStartedTourInfo className="mr-3 mb-3" />}
+                    <GettingStartedTourInfo isSourcegraphDotCom={context.isSourcegraphDotCom} className="mr-3 mb-3" />
                     <ErrorBoundary location={context.location}>
                         {objectType === 'blob' ? (
                             <>